backend/requirements.txt
Django==5.1.6
djangorestframework==3.16.0
django-cors-headers==4.7.0
channels==4.2.2
channels-redis==4.2.1
python-dotenv==1.1.0